Correct logic of decoding @0
Changes in this commit correct the output of @0 command.

Since !isNaN(0) === 1, the previous version of codes see @0 an ACommand with symbol notation like @xxx and assign it with a RAM address after RAM1023. However, @0 should be treated as an ACommand pointing to the RAM address following the @, which is RAM0.

The new codes excludes @0 when testing if the notation following @ in a ACommand is not a number, so the test will only catch those with ACommand with symbols but not @0.